[1.3.0]
### Added

- Integrated the N-Genius PayByLink feature, allowing customers to pay via a link sent to their email or phone.
