+++ Controller/NGeniusOnline/Payment.php
@@ -36,6 +36,7 @@ use NetworkInternational\NGenius\Gateway\Http\Client\TransactionFetch;
 use NetworkInternational\NGenius\Gateway\Http\TransferFactory;
 use NetworkInternational\NGenius\Gateway\Request\TokenRequest;
 use NetworkInternational\NGenius\Model\CoreFactory;
+use NetworkInternational\NGenius\Service\NgeniusApiService;
 use NetworkInternational\NGenius\Service\OrderStatusService;
 use NetworkInternational\NGenius\Setup\Patch\Data\DataPatch;
 use Ngenius\NgeniusCommon\Processor\ApiProcessor;
@@ -196,6 +197,7 @@ class Payment implements HttpGetActionInterface
      */
     private string $errorMessage = 'There is an error with the payment';
     private OrderStatusService $orderStatusService;
+    private NgeniusApiService $ngeniusApiService;
 
     /**
      * Payment constructor.
@@ -249,7 +251,8 @@ class Payment implements HttpGetActionInterface
         SerializerInterface $serializer,
         Builder $_transactionBuilder,
         OrderRepositoryInterface $orderRepository,
-        OrderStatusService $orderStatusService
+        OrderStatusService $orderStatusService,
+        NgeniusApiService $ngeniusApiService,
     ) {
         $this->request             = $request;
         $this->checkoutHelper      = $checkoutHelper;
@@ -276,6 +279,7 @@ class Payment implements HttpGetActionInterface
         $this->_transactionBuilder = $_transactionBuilder;
         $this->orderRepository     = $orderRepository;
         $this->orderStatusService  = $orderStatusService;
+        $this->ngeniusApiService = $ngeniusApiService;
     }
 
     /**
@@ -308,7 +312,7 @@ class Payment implements HttpGetActionInterface
         }
 
         if ($orderRef) {
-            $result = $this->getResponseAPI($orderRef, $storeId);
+            $result = $this->ngeniusApiService->getResponseAPI($orderRef, $storeId);
 
             $embedded = self::NGENIUS_EMBEDED;
             if ($result && isset($result[$embedded]['payment']) && is_array($result[$embedded]['payment'])) {
@@ -316,7 +320,7 @@ class Payment implements HttpGetActionInterface
 
                 $apiProcessor = new ApiProcessor($result);
                 $apiProcessor->processPaymentAction($action, $this->ngeniusState);
-                $this->processOrder($apiProcessor, $orderItem, $action);
+                $this->orderStatusService->processOrder($apiProcessor, $orderItem, $action);
             }
             if ($this->error) {
                 $this->messageManager->addError(
@@ -366,529 +370,30 @@ class Payment implements HttpGetActionInterface
         return "";
     }
 
-    /**
-     * Fetch  order details.
-     *
-     * @param string $orderRef
-     *
-     * @throws NoSuchEntityException|CouldNotSaveException
-     */
-    public function getResponseAPI($orderRef, $storeId = null): array|bool
-    {
-        $request = [
-            'token'   => $this->tokenRequest->getAccessToken($storeId),
-            'request' => [
-                'data'   => [],
-                'method' => \Laminas\Http\Request::METHOD_GET,
-                'uri'    => $this->config->getFetchRequestURL($orderRef, $storeId)
-            ]
-        ];
-        $result  = $this->transaction->placeRequest($request);
-
-        return $this->resultValidator($result);
-    }
-
-    /**
-     * Validate API response.
-     *
-     * @param array $result
-     */
-    public function resultValidator($result)
-    {
-        if (isset($result['errors']) && is_array($result['errors'])) {
-            $this->error = true;
-
-            return false;
-        } else {
-            $this->error        = false;
-            $this->ngeniusState = $result[self::NGENIUS_EMBEDED]['payment'][0]['state'] ?? '';
-
-            return $result;
-        }
-    }
-
-    /**
-     * Process Order - response from Payment Portal
-     *
-     * @param array $paymentResult
-     * @param object $orderItem
-     * @param string $orderRef
-     * @param string $action
-     *
-     * @throws CouldNotSaveException
-     * @throws InputException
-     * @throws LocalizedException
-     * @throws ValidationException
-     */
-    public function processOrder(ApiProcessor $apiProcessor, object $orderItem, string $action): void
-    {
-        $dataTable   = [];
-        $incrementId = $orderItem->getOrderId();
-
-        if ($incrementId) {
-            $paymentId = $apiProcessor->getPaymentId();
-
-            $order = $this->orderFactory->create()->loadByIncrementId($incrementId);
-
-            $storeId = $order->getStoreId();
-
-            if ($order->getStatus() !== $this->config->getCustomSuccessOrderStatus($storeId)) {
-                if ($order->getId()) {
-                    $dataTable               = $this->getCapturePayment(
-                        $order,
-                        $apiProcessor,
-                        $paymentId,
-                        $action,
-                        $dataTable
-                    );
-                    $dataTable['entity_id']  = $order->getId();
-                    $dataTable['payment_id'] = $paymentId;
-
-                    $this->updateTable($dataTable, $orderItem);
-                } else {
-                    $orderItem->setPaymentId($paymentId);
-                    $orderItem->setState($this->ngeniusState);
-                    $orderItem->setStatus($this->ngeniusState);
-                    $orderItem->save();
-                }
-            }
-        }
-    }
-
-    /**
-     * Magento order capturing
-     *
-     * @param Order $order
-     * @param array $paymentResult
-     * @param string $paymentId
-     * @param string $action
-     * @param array $dataTable
-     *
-     * @return array
-     * @throws LocalizedException
-     * @throws NoSuchEntityException
-     */
-    public function getCapturePayment(
-        Order $order,
-        ApiProcessor $apiProcessor,
-        string $paymentId,
-        string $action,
-        array $dataTable
-    ): array {
-        $paymentResult = $apiProcessor->getPaymentResult();
-        if ($apiProcessor->isPaymentConfirmed()) {
-            $order->setState(Order::STATE_PROCESSING);
-            $order->setStatus(Order::STATE_PROCESSING)->save();
-            $this->orderSender->send($order, true);
-
-            if ($action === "AUTH") {
-                $this->orderAuthorize($order, $paymentResult, $paymentId);
-            } elseif ($action === "SALE" || $action === 'PURCHASE') {
-                $dataTable['captured_amt'] = $this->orderSale($order, $paymentResult, $paymentId);
-            }
-            $dataTable['status'] = $order->getStatus();
-        } elseif ($this->ngeniusState === self::NGENIUS_STARTED) {
-            $dataTable['status'] = Order::STATE_PENDING_PAYMENT;
-        } else {
-            // Authorisation has failed - cancel order
-            // Reverse reserved stock
-            $this->error        = true;
-            $this->errorMessage = 'Result Code: ' . ($paymentResult['authResponse']['resultCode'] ?? 'FAILED')
-                . ' Reason: ' . ($paymentResult['authResponse']['resultMessage'] ?? 'Unknown');
-            $this->checkoutSession->restoreQuote();
-
-            $payment = $order->getPayment();
-
-            $formattedPrice = $order->getBaseCurrency()->formatTxt($order->getGrandTotal());
-
-            $paymentData = [
-                'Card Type'   => $paymentResult['paymentMethod']['name'] ?? '',
-                'Card Number' => $paymentResult['paymentMethod']['pan'] ?? '',
-                'Amount'      => $formattedPrice
-            ];
-
-            $trans = $this->_transactionBuilder;
-
-            $transaction = $trans->setPayment($payment)
-                                 ->setOrder($order)
-                                 ->setTransactionId($paymentId)
-                                 ->setAdditionalInformation(
-                                     [Transaction::RAW_DETAILS => $paymentData]
-                                 )
-                                 ->setFailSafe(true)
-                // Build method creates the transaction and returns the object
-                                 ->build(TransactionInterface::TYPE_CAPTURE);
-
-            $payment->addTransactionCommentsToOrder(
-                $transaction,
-                $this->errorMessage
-            );
-
-            $payment->setParentTransactionId(null);
-            $payment->save();
-            $order->save();
-
-            $transaction->save()->getTransactionId();
-            $this->updateInvoice($order, false);
-
-            $payment->setAdditionalInformation(['raw_details_info' => json_encode($paymentResult)]);
-
-            $storeId = $order->getStoreId();
-
-            if ($this->config->getCustomFailedOrderStatus($storeId) != null) {
-                $status = $this->config->getCustomFailedOrderStatus($storeId);
-            } else {
-                $status = Order::STATE_CLOSED;
-            }
-
-            if ($this->config->getCustomFailedOrderState($storeId) != null) {
-                $state = $this->config->getCustomFailedOrderState($storeId);
-            } else {
-                $state = Order::STATE_CLOSED;
-            }
-
-            $dataTable['status'] = $status;
-
-            $order->cancel()->save();
-
-            $order->setState($state);
-            $order->setStatus($status);
-            $order->save();
-
-            $order->addStatusHistoryComment('The payment on order has failed.')
-                  ->setIsCustomerNotified(false)->save();
-        }
-
-        return $dataTable;
-    }
-
-    /**
-     * Order Authorize.
-     *
-     * @param Order $order
-     * @param array $paymentResult
-     * @param string $paymentId
-     *
-     * @return null
-     * @throws Exception
-     */
-    public function orderAuthorize(Order $order, array $paymentResult, string $paymentId)
-    {
-        if ($this->ngeniusState == self::NGENIUS_AUTHORISED) {
-            $payment = $order->getPayment();
-            $payment->setLastTransId($paymentId);
-            $payment->setTransactionId($paymentId);
-            $payment->setAdditionalInformation(['paymentResult' => json_encode($paymentResult)]);
-            $payment->setIsTransactionClosed(false);
-            $formatedPrice = $order->getBaseCurrency()->formatTxt($order->getGrandTotal());
-
-            $paymentData = [
-                'Card Type'   => $paymentResult['paymentMethod']['name'] ?? '',
-                'Card Number' => $paymentResult['paymentMethod']['pan'] ?? '',
-                'Amount'      => $formatedPrice
-            ];
-
-            $transactionBuilder = $this->transactionBuilder->setPayment($payment)
-                                                           ->setOrder($order)
-                                                           ->setTransactionId($paymentId)
-                                                           ->setAdditionalInformation(
-                                                               [Transaction::RAW_DETAILS => $paymentData]
-                                                           )->setAdditionalInformation(
-                                                               ['paymentResult' => json_encode($paymentResult)]
-                                                           )
-                                                           ->setFailSafe(true)
-                                                           ->build(
-                                                               Transaction::TYPE_AUTH
-                                                           );
-
-            $payment->addTransactionCommentsToOrder($transactionBuilder, null);
-            $payment->setParentTransactionId(null);
-            $payment->save();
-
-            $message = 'The payment has been approved and the authorized amount is ' . $formatedPrice;
-
-            $status = Order::STATE_PROCESSING;
-
-            $this->updateOrderStatus($order, $status, $message);
-        }
-    }
-
-    /**
-     * Order Status Updater
-     *
-     * @param Order $order
-     * @param ?string $status
-     * @param string $message
-     *
-     * @return void
-     * @throws NoSuchEntityException
-     */
-    private function updateOrderStatus(Order $order, ?string $status, string $message): void
-    {
-        //Check For Custom Order Status on Payment Complete
-        $storeId = $order->getStoreId();
-
-        if ($this->config->getCustomSuccessOrderStatus($storeId) != null) {
-            $status = $this->config->getCustomSuccessOrderStatus($storeId);
-        }
-
-        if ($this->config->getCustomSuccessOrderState($storeId) != null) {
-            $order->setState($this->config->getCustomSuccessOrderState($storeId));
-        }
-
-        $order->addStatusToHistory($status, $message, true);
-        $order->save();
-    }
-
-    /**
-     * Order Sale.
-     *
-     * @param object $order
-     * @param array $paymentResult
-     * @param string $paymentId
-     *
-     * @return null|float
-     * @throws Exception
-     */
-    public function orderSale($order, $paymentResult, $paymentId)
-    {
-        if ($this->ngeniusState === self::NGENIUS_CAPTURED || $this->ngeniusState === self::NGENIUS_PURCHASED) {
-            $payment = $order->getPayment();
-            $payment->setLastTransId($paymentId);
-            $payment->setTransactionId($paymentId);
-            $payment->setAdditionalInformation(['paymentResult' => json_encode($paymentResult)]);
-            $payment->setIsTransactionClosed(false);
-            $grandTotal    = $order->getGrandTotal();
-            $formatedPrice = $order->getBaseCurrency()->formatTxt($grandTotal);
-
-            $paymentData = [
-                'Card Type'   => $paymentResult['paymentMethod']['name'] ?? '',
-                'Card Number' => $paymentResult['paymentMethod']['pan'] ?? '',
-                'Amount'      => $formatedPrice
-            ];
-
-            $transactionId = $paymentResult['reference'];
-
-            $transactionBuilder = $this->transactionBuilder->setPayment($payment)
-                                                           ->setOrder($order)
-                                                           ->setTransactionId($transactionId)
-                                                           ->setAdditionalInformation(
-                                                               [Transaction::RAW_DETAILS => (array)$paymentData]
-                                                           )
-                                                           ->setAdditionalInformation(
-                                                               ['paymentResult' => json_encode($paymentResult)]
-                                                           )
-                                                           ->setFailSafe(true)
-                                                           ->build(
-                                                               Transaction::TYPE_CAPTURE
-                                                           );
-
-            $payment->addTransactionCommentsToOrder($transactionBuilder, null);
-            $payment->setParentTransactionId(null);
-            $payment->save();
-
-            $message = 'The payment has been approved and the captured amount is ' . $formatedPrice;
-
-            if ($order->canShip()) {
-                $status = Order::STATE_PROCESSING;
-            } else {
-                $status = Order::STATE_COMPLETE;
-            }
-
-            $this->updateOrderStatus($order, $status, $message);
-
-            $this->updateInvoice($order, true, $transactionId);
-
-            return $grandTotal;
-        }
-    }
-
-    /**
-     * Update Invoice.
-     *
-     * @param object $order
-     * @param bool $flag
-     * @param string $transactionId
-     *
-     * @return null
-     * @throws LocalizedException
-     * @throws Exception
-     */
-    public function updateInvoice($order, $flag, $transactionId = null)
-    {
-        if ($order->hasInvoices()) {
-            // gets here from a 'SALE' transaction
-            if ($flag === false) {
-                foreach ($order->getInvoiceCollection() as $invoice) {
-                    $invoice->cancel()->save();
-                }
-            } else {
-                foreach ($order->getInvoiceCollection() as $invoice) {
-                    $this->doUpdateInvoice($invoice, $transactionId, $order);
-                }
-            }
-        } elseif ($flag) {
-            // Create invoice - gets here from a 'PURCHASE' transaction
-            $invoice = $this->invoiceService->prepareInvoice($order);
-            $invoice->register();
-            $payment = $order->getPayment();
-            $payment->setCreatedInvoice($invoice);
-            $order->setPayment($payment);
-            $this->doUpdateInvoice($invoice, $transactionId, $order);
-        }
-    }
-
-    /**
-     * Invoice Updater
-     *
-     * @param InvoiceInterface|Invoice $invoice
-     * @param string|null $transactionId
-     * @param object $order
-     *
-     * @return void
-     * @throws Exception
-     */
-    public function doUpdateInvoice(
-        InvoiceInterface|Invoice $invoice,
-        ?string $transactionId,
-        object $order
-    ): void {
-        $invoice->setRequestedCaptureCase(Invoice::CAPTURE_ONLINE);
-        $invoice->setTransactionId($transactionId);
-        $invoice->pay()->save();
-        $transactionSave = $this->transactionFactory->create()->addObject($invoice)->addObject(
-            $invoice->getOrder()
-        );
-        $transactionSave->save();
-
-        if ($this->config->getInvoiceSend()) {
-            try {
-                $this->invoiceSender->send($invoice);
-                $order->addStatusHistoryComment(
-                    __('Notified the customer about invoice #%1.', $invoice->getIncrementId())
-                )
-                      ->setIsCustomerNotified(true)->save();
-            } catch (Exception $e) {
-                $this->messageManager->addError(__('We can\'t send the invoice email right now.'));
-            }
-        }
-    }
-
-    /**
-     * Update Table.
-     *
-     * @param array $data
-     * @param object $orderItem
-     *
-     * @return bool true
-     */
-    public function updateTable(array $data, $orderItem)
-    {
-        $orderItem->setEntityId($data['entity_id']);
-        $orderItem->setState($this->ngeniusState);
-        $orderItem->setStatus($data['status']);
-        $orderItem->setPaymentId($data['payment_id']);
-        if (isset($data['captured_amt'])) {
-            $orderItem->setCapturedAmt($data['captured_amt']);
-        }
-        $orderItem->save();
-
-        return true;
-    }
 
     /**
      * Cron Task.
      */
     public function cronTask(): void
     {
-        $orderItems = $this->fetchOrder('state', self::NGENIUS_STARTED)->addFieldToFilter(
-            'payment_id',
-            null
-        )->addFieldToFilter('created_at', ['lteq' => date('Y-m-d H:i:s', strtotime('-1 hour'))])->setOrder(
-            'nid',
-            'DESC'
-        );
+        $orderItems = $this->fetchOrder('state', self::NGENIUS_STARTED)
+            ->addFieldToFilter('payment_id', null)
+            ->addFieldToFilter('created_at', ['lteq' => date('Y-m-d H:i:s', strtotime('-1 hour'))])
+            ->setOrder('nid', 'DESC');
 
         $pblOrderItems = $this->fetchOrder('state', $this->orderStatusService->getDefaultPBLState())
-                              ->addFieldToFilter('payment_id', null)
-                              ->addFieldToFilter('created_at', ['lteq' => date('Y-m-d H:i:s', strtotime('-3 days'))])
-                              ->setOrder('nid', 'DESC');
+            ->addFieldToFilter('payment_id', null)
+            ->addFieldToFilter('created_at', ['lteq' => date('Y-m-d H:i:s', strtotime('-1 hour'))])
+            ->setOrder('nid', 'DESC');
 
-        $orderItems = array_merge(
-            $orderItems->getItems(),
-            $pblOrderItems->getItems()
-        );
+        if ($orderItems->getItems()) {
+            $this->logger->info("N-GENIUS: Found " . count($orderItems->getItems()) . " unprocessed normal order(s)");
+            $this->orderStatusService->processNormalOrders($orderItems->getItems());
+        }
 
-        if ($orderItems) {
-            $this->logger->info("N-GENIUS: Found " . count($orderItems) . " unprocessed order(s)");
-            $counter = 0;
-            foreach ($orderItems as $orderItem) {
-                if ($counter >= 5) {
-                    $this->logger->info("N-GENIUS: Breaking loop at 5 orders to avoid timeout");
-                    break;
-                }
-
-                $orderItem->setState('cron');
-                $orderItem->setStatus('cron');
-                $orderItem->save();
-
-                $orderRef    = $orderItem->getReference();
-                $incrementId = $orderItem->getOrderId();
-
-                $order = $this->orderFactory->create()->loadByIncrementId($incrementId);
-
-                if (!$order->getId()) {
-                    $this->logger->info("N-GENIUS: Magento order not found");
-                    break;
-                }
-
-                try {
-                    if ($order->getPayment()->getMethod() !== 'ngeniusonline') {
-                        $this->logger->info("N-GENIUS: Order#" . $incrementId . " will not be processed");
-                        continue;
-                    }
-
-                    $this->logger->info("N-GENIUS: Processing order $incrementId");
-
-                    $order->addStatusHistoryComment(
-                        __('This is order is being processed by the cron.')
-                    )->save();
-
-                    $storeId = $order->getStoreId();
-
-                    $result   = $this->getResponseAPI($orderRef, $storeId);
-                    $embedded = self::NGENIUS_EMBEDED;
-                    if ($result && isset($result[$embedded]['payment']) && is_array($result[$embedded]['payment'])) {
-                        $action = $result['action'] ?? '';
-
-                        $apiProcessor  = new ApiProcessor($result);
-                        $paymentResult = $apiProcessor->getPaymentResult();
-                        $this->logger->info('N-GENIUS: state is ' . $paymentResult['state']);
-                        if ($paymentResult['state'] === self::NGENIUS_STARTED
-                            || $paymentResult['state'] === self::NGENIUS_AWAIT3DS
-                            || $paymentResult['state'] == self::NGENIUS_PENDING
-                            || $paymentResult['state'] == self::NGENIUS_CANCELLED
-                        ) {
-                            $paymentResult['state'] = "FAILED";
-                            $this->ngeniusState     = self::NGENIUS_FAILED;
-                        }
-                        $this->processOrder($apiProcessor, $orderItem, $action);
-                    } else {
-                        $this->logger->info("N-GENIUS: Payment result not found");
-                        $order->addStatusHistoryComment(
-                            __('N-GENIUS Payment result not found.')
-                        )->save();
-                        $this->logger->info("N-GENIUS: Result " . json_encode($result));
-                    }
-
-                    $counter++;
-                } catch (Exception $e) {
-                    $this->logger->info('N-GENIUS: exception ' . $e->getMessage());
-                    $order->addStatusHistoryComment(
-                        __('N-GENIUS: Exception ' . $e->getMessage())
-                    )->save();
-                }
-            }
+        if ($pblOrderItems->getItems()) {
+            $this->logger->info("N-GENIUS: Found " . count($pblOrderItems->getItems()) . " unprocessed PBL order(s)");
+            $this->orderStatusService->processPBLOrders($pblOrderItems->getItems());
         }
     }
 }
