+++ Helper/Version.php
@@ -4,7 +4,7 @@ namespace NetworkInternational\NGenius\Helper;
 
 class Version
 {
-    const MODULE_VERSION = '1.2.0';
+    const MODULE_VERSION = '1.3.0';
 
     public function getVersion()
     {
