+++ composer.json
@@ -2,7 +2,7 @@
   "name": "networkinternational/ngenius",
   "description": "N-Genius Online by Network Payment Gateway for Magento 2",
   "type": "magento2-module",
-  "version": "1.2.0",
+  "version": "1.3.0",
   "minimum-stability": "stable",
   "license": "GPL-3.0",
   "authors": [
