+++ etc/module.xml
@@ -1,6 +1,6 @@
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
         xsi:noNamespaceSchemaLocation="urn:magento:framework:Module/etc/module.xsd">
-    <module name="NetworkInternational_NGenius" setup_version="1.2.0">
+    <module name="NetworkInternational_NGenius" setup_version="1.3.0">
         <sequence>
             <module name="Magento_Sales"/>
             <module name="Magento_Payment"/>
